# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/themes": {
      "get": {
        "description": "Devuelve el catálogo global más los temas propios de la organización, en ese orden. Solo pide membresía activa.",
        "operationId": "ThemesController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo global seguido de los temas propios.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                      },
                      "name": {
                        "type": "string"
                      },
                      "code": {
                        "nullable": true,
                        "type": "string"
                      },
                      "source": {
                        "type": "string",
                        "enum": [
                          "catalog",
                          "organization"
                        ]
                      },
                      "editable": {
                        "type": "boolean"
                      },
                      "tokens": {},
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "code",
                      "source",
                      "editable",
                      "tokens",
                      "createdAt",
                      "updatedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Listar temas disponibles",
        "tags": [
          "themes"
        ]
      },
      "post": {
        "description": "Requiere el permiso `theme.manage` (OWNER/ADMIN). Los tokens se validan en el servidor, incluida la comprobación de contraste WCAG 2.2 AA.",
        "operationId": "ThemesController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "tokens": {}
                },
                "required": [
                  "name",
                  "tokens"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tema creado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "nullable": true,
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "catalog",
                        "organization"
                      ]
                    },
                    "editable": {
                      "type": "boolean"
                    },
                    "tokens": {},
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "code",
                    "source",
                    "editable",
                    "tokens",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "422": {
            "description": "Los tokens no cumplen el esquema o el contraste exigido. `issues` indica el campo (`palette.mutedForeground`, …)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Crear un tema propio",
        "tags": [
          "themes"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/themes/{themeId}": {
      "get": {
        "operationId": "ThemesController_getOne",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "path",
            "description": "Tema del catálogo global o propio de la organización.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El tema solicitado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "nullable": true,
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "catalog",
                        "organization"
                      ]
                    },
                    "editable": {
                      "type": "boolean"
                    },
                    "tokens": {},
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "code",
                    "source",
                    "editable",
                    "tokens",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Tema no encontrado, o pertenece a otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Leer un tema",
        "tags": [
          "themes"
        ]
      },
      "patch": {
        "description": "Los tokens se envían completos, nunca a medias: un parche parcial no se puede verificar contra AA.",
        "operationId": "ThemesController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "path",
            "description": "Tema propio de la organización.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  },
                  "tokens": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tema actualizado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "nullable": true,
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "catalog",
                        "organization"
                      ]
                    },
                    "editable": {
                      "type": "boolean"
                    },
                    "tokens": {},
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "code",
                    "source",
                    "editable",
                    "tokens",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "El tema es del catálogo global: no se edita, se duplica."
          },
          "404": {
            "description": "Tema no encontrado, o pertenece a otra organización (ADR-002)."
          },
          "422": {
            "description": "Los tokens no cumplen el esquema o el contraste exigido."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Editar un tema propio",
        "tags": [
          "themes"
        ]
      },
      "delete": {
        "description": "Solo si ningún sitio lo tiene aplicado: borrarlo cambiaría la apariencia de esos sitios en silencio.",
        "operationId": "ThemesController_remove",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "path",
            "description": "Tema propio de la organización.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tema eliminado."
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "El tema es del catálogo global: no se elimina."
          },
          "404": {
            "description": "Tema no encontrado, o pertenece a otra organización (ADR-002)."
          },
          "409": {
            "description": "Algún sitio tiene el tema aplicado. Cámbialos de tema antes de eliminarlo."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Eliminar un tema propio",
        "tags": [
          "themes"
        ]
      }
    },
    "/api/v1/organizations/{organizationId}/themes/{themeId}/duplicate": {
      "post": {
        "description": "Único camino para personalizar un tema del catálogo, que es de solo lectura. La copia queda como tema propio y editable.",
        "operationId": "ThemesController_duplicate",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "Organización dueña del recurso. Resuelve el tenant antes de tocar nada (ADR-002).",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "themeId",
            "required": true,
            "in": "path",
            "description": "Tema a copiar: del catálogo o propio.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-Requested-With",
            "in": "header",
            "required": false,
            "description": "Obligatoria con valor `impulza-one` en toda petición que modifica estado. Ver CsrfGuard.",
            "schema": {
              "type": "string",
              "enum": [
                "impulza-one"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 60
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Copia creada, editable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                    },
                    "name": {
                      "type": "string"
                    },
                    "code": {
                      "nullable": true,
                      "type": "string"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "catalog",
                        "organization"
                      ]
                    },
                    "editable": {
                      "type": "boolean"
                    },
                    "tokens": {},
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "code",
                    "source",
                    "editable",
                    "tokens",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (detalle en `issues`)."
          },
          "401": {
            "description": "Sin sesión válida."
          },
          "403": {
            "description": "Sin membresía activa en la organización, sin el permiso requerido, o petición sin la cabecera anti-CSRF."
          },
          "404": {
            "description": "Tema no encontrado, o pertenece a otra organización (ADR-002)."
          }
        },
        "security": [
          {
            "sesion": []
          }
        ],
        "summary": "Duplicar un tema",
        "tags": [
          "themes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Impulza One API",
    "description": "API REST de Impulza One. Todo cuelga de `/api/v1` salvo `/health`, que es un endpoint de\ninfraestructura sin sesión ni CSRF.\n\n**Autenticación**: cookie de sesión `HttpOnly` (`impulza_session`), emitida por\n`POST /auth/login`. No hay tokens en el cuerpo ni en cabeceras de autorización.\n\n**CSRF**: toda petición que modifica estado (POST, PATCH, PUT, DELETE) debe enviar la\ncabecera `X-Requested-With: impulza-one`. Un formulario o una imagen entre sitios no puede\nfijar cabeceras propias, y una petición `fetch` con cabecera custom dispara preflight CORS,\nbloqueado salvo origen autorizado. Sin ella la respuesta es 403.\n\n**Multi-tenant**: los recursos de negocio cuelgan de `/organizations/:organizationId/…`.\nUn recurso de otra organización responde **404 y no 403**, para no confirmar que existe\n(ADR-002).\n\n**Errores de validación**: 400 cuando el cuerpo no cumple el esquema de entrada, 422 cuando\ncumple la forma pero viola una regla de dominio (contraste insuficiente en un tema,\nconfiguración inválida para un tipo de bloque). Ambos traen `issues` con el campo exacto.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "apiKey",
        "in": "cookie",
        "name": "impulza_session"
      }
    },
...
  }
}